pyproject.toml
[tool.poetry]
name = "google-timeline-visualiser"
version = "0.1.0"
description = "Analysis and visualization of Google Timeline location data"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
googlemaps = "^4.10.0"
numpy = "^1.26.0"
pandas = "^2.2.0"
scikit-learn = "^1.3.0"
folium = "^0.14.0"
matplotlib = "^3.7.0"
python-dotenv = "^1.0.0"
icecream = "^2.1.3"
seaborn = "^0.13.2"
geopandas = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 